Tighten chat viewport and image handling
