+++ macos/Podfile
@@ -38,5 +38,11 @@ end
 post_install do |installer|
   installer.pods_project.targets.each do |target|
     flutter_additional_macos_build_settings(target)
+    target.build_configurations.each do |config|
+      next unless config.name == 'Profile'
+
+      # ogg_opus_player ships an arm64-only static library for macOS.
+      config.build_settings['ARCHS'] = 'arm64'
+    end
   end
 end
